API module and middleware overhaul
