+++ api/dataProcessing/types.ts
@@ -1,0 +1,13 @@
+export type Column = {
+  original_column: string;
+  sql_column: string;
+};
+
+export type AlertRecord = {
+  territory_name: string;
+  alert_type: string;
+  month_detec: string;
+  year_detec: string;
+  area_alert_ha: string;
+  _topic: string;
+}
